- Nokia E71's default browser cannot handle a 0 byte image.  It does not fire any events, so we use a 35 byte gif for the latency test instead.  A 1x1 gif is smaller than a 1x1 png.
- When an image times out, we were jumping to the next run immediately, however, the browser
still downloads the image in the background, which affects the results of the next run.  If the
next run is a latency check, it could mean that the latency test never runs because of the
parallel downloads from same host restriction.

To get around this, we now wait until a timed out image has downloaded, and use its values anyway,
but do not proceed to the next image once the onload/onerror event fires